requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.0.0
scikit-learn>=1.2.0
numpy>=1.21.0
pandas>=1.5.0
requests>=2.28.0
python-dotenv>=1.0.0
networkx>=2.8
google-generativeai>=0.3.0
